sync

+++ task_8/src/aux/func.py
@@ -1,9 +1,8 @@
 import numpy as np
 
 from numba import njit
-
 from aux.var import A, B
 
 @njit
 def func(t, y):
-    return np.array([A * y[1], -B * y[0]])
+    return np.array([A * y[1], -B * y[0]], dtype=np.float32)
